Update to ebookmaker 0.11.26
Recent versions of ebookmaker have used a config file to configure Tidy
which is a required part of the ebookmaker process. Version 0.11.26 has
a command line argument to specify the location of the config file, since
otherwise it was hidden in the virtualenv.

1. Update pipfile
2. Script get_tidy_conf.sh gets the tidy.conf for a given ebookmaker version
and copies it to the dist folder ready for release
3. Update documentation
4. Sample book now contains `nbsp` characters to test tidy.conf mechanism